+++ pjmedia/src/pjmedia-codec/vpx.c
@@ -830,7 +830,7 @@ on_return:
 
         PJ_LOG(4,(THIS_FILE, "Decode couldn't produce picture, "
                   "input nframes=%lu, concatenated size=%d bytes",
-                  count, whole_len));
+                  (unsigned long)count, whole_len));
 
         output->type = PJMEDIA_FRAME_TYPE_NONE;
         output->size = 0;
